send minimal info for process data
add wmstats proc info
