+++ src/python/WMCore/ReqMgr/Service/Auxiliary.py
@@ -10,7 +10,6 @@ import json
 import cherrypy
 import time
 
-from Utils.ProcessStats import processStatus, threadStack
 import WMCore
 from WMCore.Database.CMSCouch import Document, CouchNotFoundError, CouchError
 from WMCore.REST.Error import RESTError, NoSuchInstance, APINotSpecified
@@ -80,23 +79,6 @@ class Info(RESTEntity):
         # "return result" only would return dict with keys without (!!) values set
         return rows([result])
 
-class ProcessMatrix(RESTEntity):
-    """
-    Process and thread matrix in reqmgr2
-    """
-    def __init__(self, app, api, config, mount):
-        # CouchDB auxiliary database name
-        RESTEntity.__init__(self, app, api, config, mount)
-        self.time0 = time.time()
-
-    def validate(self, apiobj, method, api, param, safe):
-        pass
-
-    @restcall(formats=[('text/plain', PrettyJSONFormat()), ('application/json', JSONFormat())])
-    def get(self):
-        sdict = {'server': processStatus()}
-        sdict['server'].update({'uptime': time.time() - self.time0})
-        return rows([sdict])
 
 class ReqMgrConfigData(RESTEntity):
     def __init__(self, app, api, config, mount):
